refactor features directory.
